feat: modularize feature of showing total income
